requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.8
openai>=1.3.0
chromadb>=0.4.6
numpy>=1.24.0
pandas>=2.0.0
pymysql>=1.1.0
pytest>=7.3.1
httpx>=0.24.0
jinja2>=3.1.2
pyyaml>=6.0
typing-extensions>=4.5.0
sentence-transformers>=2.2.2
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
greenlet>=3.0.0 